revamped stats to preserve dynamic data
Previously, only the VecFieldCartesian instances were saved with the
MeanVecFieldCartesian parent instance, which was really kind of a sloppy
way to preserve that information. Turns out we are interested in
performing more complex statistics in the time domain on the fluctuating
components, so this information needs to be tracked in a manner which is
more consistent with the vel_matrix. Its worth noting that dynamic_set
and vel_matrix are not very well named. it should probably be
dynamic_set and mean_set.