extensive addition of automatic plot generation
Added several classes and a new module called texmanager, this code can
programatically produce figures and create tex files for use in the
appendices of the thesis document. example plots are still produced by
the old method, which should probably be revamped since its a bit
raviolli at the moment.