giving up
i know its possible to make this perform better, but i cant figure it
out.
One should be able to use the reverse transforms to convert a particles
known x and y coordinates in millimeters back into pixel space, which
should correspond to the indices of the matrices and allow some
efficient sub assignments to the intensity matrix for each particle.
Sadly, these values always only work out to be approximate, they are
incorrect somehow, and I am not sure if this is symptomatic of a larger
problem.