More changes to uncertainty image scheme
having 4 calibration sets for extreme cases seems to be a bit confusing,
and is perhaps a little too harsh. Strategy up until this point has been
to always assume the worst case scenario. Well, it turns out, when you
make lots of worst case assumptions, you end up showing the data is
worthless. The data doesn't appear to be worthless, so I'm going to test
an approach where the monte carlo images represent the typical
velocities seen in a test, rather than the extremes. Results in fewer
tests, but also hopefully in better representative values.