redo uncertainty images one more time
The approach used previously was the right approach, it was just based
on the wrong numbers. The old table took absolute maximum velocity
values, rather than maximums along the best fit curves. this made them
extremely succeptible to noise. This has been fixed, but now the
azimuthal velocity ranges to simulate have been reduced dramatically.
Axial velocities are being placed back into the center of the range,
since minimum axial velocity at the core is not actually of interest as
presumed.

texdocs: the section on PIV uncertainty has been so strongly influenced
by results, that it has been moved to the results section of the texdoc.